added api message options
In case the API message contains option of type deprecated
or in-progress, the information is displayed in the generated
code in a form of additional comment:

// InProgress: <info>
// Deprecated: <info>

In case the <info> is not provided, a generic message is shown.
Possible future use to automatically search whether such
messages are in use.

Change-Id: Icf87cc9a2fe6bf31f7555320255c9f0736add6e1
Signed-off-by: Vladimir Lavor <vlavor@cisco.com>