statseg2: robust symlink loop detections
Symlink loop detection relied on the range of indices being copied out.
Encountered a symlink, the function recursed, calling itself with the
range of size 1 (value extracted from the symlink).  Symlink referring
to a symlink is not allowed.  Whenever the function encountered a
symlink when called with a reduced range of indices, it bailed out due
to supposed symlink loop.

However the range could be reduced for other reasons (dump limit
feature).  Introduce a helper function and an explicit 'nofollow'
parameter.