Add GetRetVal fonction to ReplyMessage interface
This patch adds a GetRetVal function on all reply
messages, that retreives the error value VPP
returns in its Reply. If the field is not present
or the value is 0 (success) it returns nil.

This also changes the types enum to xxxType names
e.g. api.RequestMessage becomes api.RequestMessageType

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>
Change-Id: I55ca4d02b19909a3fd44c152962b934621646e9f