Seperate version constants into *.version.go to ease churn on regen
Currently the version constant stanzas like:

const (
	APIFile    = "stn"
	APIVersion = "2.0.0"
	VersionCrc = 0x9cfaef64
)

are generated in *.ba.go.  This causes those files to be replaced
on regen, even if nothing else is changed.

This PR puts that stanza into a *.version.go file, so that someone
can tell at a glance whether or not there are meaningful updates
to generated API files.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>
Change-Id: I8fc2c0dfa70f078d6b7f916fb800cd0f97e0b13c