linter: move to staticcheck
As golint is now deprecated, this patch proposes to
switch to staticcheck [0] as an alternative, in addition
to `go vet`.
This patch contains fixes addressing the linter's reported
issues, with two added exceptions :
- The MemoryStats deprectated fields are still for us to update
- In `gen_encoding.go` we need to add an extra comment
to account for the unused variable. This is probably worth
addressing in another patch to not generate a variable we don't use.

[0] https://staticcheck.io/

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>
Change-Id: I0562f290fb908bc610c44d675e21e466d0badc27