Introducing ChannelProvidingConnection
api.Connection is a very convenient interface for using
the RPC style binapi

Unfortunately, in order to Subscribe to Events like
interfaces.SwInterfaceEvent{} you have to have an api.Channel
against which you can call SubscribeNotification

Conveniently, core.Connection implements both api.Connection and
api.ChannelProvider.

Unfortunately,  we have no type with which a function can indicate
it requires *both* of those interfaces.

ChannelProvidingConnection implements both api.ChannelProvider and
api.Connection allowing it to be used as a parameter to functions
which require an argument that implements both.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>