Fix channel pool
This change fixes the issue with channel pool causing the IDs be exhausted due to the pool deallocating the channels in the pool. To resolve this there is a new pool for channel IDs which is used to cache unused channel IDs.
There are some minor changes in debug logs.

Signed-off-by: Ondrej Fabry <ofabry@cisco.com>