fix handling of EAGAIN reply return value
The new VPP binapi for stream can return a reply with the return value EAGAIN. This value signals, that VPP ends the transmition because the call has taken to long and is possibly blocking the master thread.
The client is then expected to repeat the request, optimally providing the cursor to the last received data.
Sofar, the implementation only returned an error without providing the cursor to the last received data point (in older versions, e.g., v0.4.0, govpp does not return an error and just accepts the incomplete response).